backend1/requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
